requirements.txt
dnspython>=2.4.0
requests>=2.31.0
colorama>=0.4.6
rich>=13.0.0
cryptography>=41.0.0